Added additional testcases for better coverage
